Adding ReadmeFiles
Adding zip archives of files created that were not under the scope of the repository.
The JenkinsBuild.zip includes a working version of a build produced by the JenkinsBuild machine I set up.
It also contains a Results.xml file, which is a report of the unit tests, run and output while the Jenkins is making a new build.
The second is TechnicalTestRPS.7z, which is an archive of the JenkinsBuild project set up on my local machine, which acted as the localhost.